ocr_server/requirements.txt
accelerate==1.3.0
gradio==5.14.0
pillow==11.1.0
requests==2.32.3
sentencepiece==0.2.0
torch==2.6.0
torchvision==0.21.0
transformers==4.48.2
flash-attn==2.7.4.post1
git+https://github.com/deepseek-ai/Janus.git